Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-cors-exposed-header-names.https-expected.txt



FAIL CORS-exposed header names for a response from sw assert_equals: expected (string) "bar" but got (object) null
